refactor(rsources): simplification pass over the delegate
Component files 5->3 (config subscription and stubs fold into the main file), the fail-loud constructor unexported, the test probe shares the real constructor's field initialisation instead of hand-rebuilding it, factories fail at construction on a missing SyncSettings, and behavior pinned twice at different layers keeps one pin. Net -135 lines, no design or metric changes.
