+++ services/rsources/sync_setting_delegate.go
@@ -17,6 +17,7 @@ import (
 
 	backendconfig "github.com/rudderlabs/rudder-server/backend-config"
 	"github.com/rudderlabs/rudder-server/jobsdb"
+	"github.com/rudderlabs/rudder-server/utils/misc"
 	"github.com/rudderlabs/rudder-server/utils/pubsub"
 )
 
@@ -82,9 +83,7 @@ type syncSettingDelegate struct {
 	conf        *config.Config
 	db          *sql.DB
 
-	// Reloadable knobs, read with .Load() at the point of use. Deliberately not
-	// snapshotted per batch: a snapshot buys uniformity within one status-update
-	// batch, which nothing depends on, at the cost of a struct copy on a hot path.
+	// Reloadable knobs, read with .Load() at the point of use.
 	enabled            config.ValueLoader[bool]
 	maxErrorLength     config.ValueLoader[int]
 	blockedConnections config.ValueLoader[[]string]
@@ -149,6 +148,35 @@ type suppressionKey struct {
 
 var _ SyncSettingDelegate = (*syncSettingDelegate)(nil)
 
+// newDelegate builds the in-memory half of the component: the config bindings, the
+// caches and the shutdown plumbing, over an already-open database handle.
+//
+// Split from NewSyncSettingDelegate so that tests wanting a delegate over a fake
+// database get the real field wiring instead of hand-rebuilding it - a copy that
+// silently drifts the moment a field is added here.
+func newDelegate(conf *config.Config, log logger.Logger, statFactory stats.Stats, db *sql.DB) *syncSettingDelegate {
+	if log == nil {
+		log = logger.NOP
+	}
+	return &syncSettingDelegate{
+		log:                log,
+		statFactory:        statFactory,
+		conf:               conf,
+		db:                 db,
+		enabled:            conf.GetReloadableBoolVar(false, captureErrorDetailKey),
+		maxErrorLength:     conf.GetReloadableIntVar(defaultMaxErrorLength, 1, maxErrorLengthKey),
+		blockedConnections: conf.GetReloadableStringSliceVar(nil, blockedConnectionsKey),
+		blockedWorkspaces:  conf.GetReloadableStringSliceVar(nil, blockedWorkspacesKey),
+		maxAge:             conf.GetReloadableDurationVar(24, time.Hour, syncSettingsMaxAgeKey),
+		cleanupFrequency:   conf.GetReloadableDurationVar(1, time.Hour, syncSettingsCleanupFrequencyKey),
+		decisions:          make(map[string]syncSettingEntry),
+		connections:        make(map[connectionKey]bool),
+		configLoaded:       make(chan struct{}),
+		warned:             make(map[suppressionKey]struct{}),
+		stop:               make(chan struct{}),
+	}
+}
+
 // NewSyncSettingDelegate builds the component: it opens its own database connection,
 // creates its table if needed, loads every pinned decision into memory and subscribes
 // to the backend config.
@@ -168,30 +196,11 @@ func NewSyncSettingDelegate(
 	statFactory stats.Stats,
 	backendConfig backendconfig.BackendConfig,
 ) (*syncSettingDelegate, error) {
-	if log == nil {
-		log = logger.NOP
-	}
 	db, err := setupSyncSettingsDBConn(conf, statFactory)
 	if err != nil {
 		return nil, fmt.Errorf("rsources sync settings: db setup: %w", err)
 	}
-	d := &syncSettingDelegate{
-		log:                log,
-		statFactory:        statFactory,
-		conf:               conf,
-		db:                 db,
-		enabled:            conf.GetReloadableBoolVar(false, captureErrorDetailKey),
-		maxErrorLength:     conf.GetReloadableIntVar(defaultMaxErrorLength, 1, maxErrorLengthKey),
-		blockedConnections: conf.GetReloadableStringSliceVar(nil, blockedConnectionsKey),
-		blockedWorkspaces:  conf.GetReloadableStringSliceVar(nil, blockedWorkspacesKey),
-		maxAge:             conf.GetReloadableDurationVar(24, time.Hour, syncSettingsMaxAgeKey),
-		cleanupFrequency:   conf.GetReloadableDurationVar(1, time.Hour, syncSettingsCleanupFrequencyKey),
-		decisions:          make(map[string]syncSettingEntry),
-		connections:        make(map[connectionKey]bool),
-		configLoaded:       make(chan struct{}),
-		warned:             make(map[suppressionKey]struct{}),
-		stop:               make(chan struct{}),
-	}
+	d := newDelegate(conf, log, statFactory, db)
 	if err := d.setupTable(ctx); err != nil {
 		_ = db.Close()
 		return nil, fmt.Errorf("rsources sync settings: %w", err)
@@ -241,8 +250,11 @@ func (d *syncSettingDelegate) GetErrorResponse(ctx context.Context, key statKey,
 	return text, nil
 }
 
-// emptyErrorResponse reports whether the raw envelope holds nothing at all. `”` and
-// `{}` are the two shapes the pipeline writes for "no detail recorded"; surrounding
+// emptyErrorResponse reports whether the raw envelope holds nothing at all.
+//
+// Empty, `”` and `{}` are the three shapes enumerated by the review that specified
+// this check order; no producer of the `”` form exists in this repository, so it is
+// carried on the reviewer's authority rather than on an observed writer. Surrounding
 // whitespace is tolerated so that a `{}\n` shaped value short-circuits too.
 func emptyErrorResponse(errorResponse json.RawMessage) bool {
 	trimmed := bytes.TrimSpace(errorResponse)
@@ -359,4 +371,141 @@ func (d *syncSettingDelegate) stopping() bool {
 	default:
 		return false
 	}
+}
+
+// ConfigSubscriberRoutine consumes the component's backend-config subscription and
+// rebuilds the connection index on every push.
+//
+// It returns when the subscription closes (utils/pubsub closes it when the context
+// handed to Subscribe - the constructor's - is done), when ctx is cancelled, or when
+// Stop is called. The last of those matters: without it, a Stop that does not also
+// cancel the constructor's context would block forever in wg.Wait().
+func (d *syncSettingDelegate) ConfigSubscriberRoutine(ctx context.Context) error {
+	if !d.routineStarted() {
+		return nil
+	}
+	defer d.wg.Done()
+	for {
+		select {
+		case <-ctx.Done():
+			return nil
+		case <-d.stop:
+			return nil
+		case data, open := <-d.configCh:
+			if !open {
+				return nil
+			}
+			workspaces, ok := data.Data.(map[string]backendconfig.ConfigT)
+			if !ok {
+				// TopicProcessConfig always carries map[string]ConfigT; a different
+				// shape means the topic contract changed under us and every decision
+				// from here on would silently be false.
+				d.log.Errorn("rsources sync settings: unexpected backend config payload")
+				continue
+			}
+			d.indexConnections(workspaces)
+		}
+	}
+}
+
+// indexConnections rebuilds the {sourceID, destinationID} -> enabled index wholesale.
+//
+// Rebuilt rather than merged so that a connection removed from the control plane
+// disappears here too; a partial update would leave a deleted connection answering
+// `true` forever. The index spans every workspace in the push - one process serves
+// many workspaces, and the connection identity is already globally unique.
+func (d *syncSettingDelegate) indexConnections(workspaces map[string]backendconfig.ConfigT) {
+	index := make(map[connectionKey]bool)
+	for _, wConfig := range workspaces {
+		for _, conn := range wConfig.Connections {
+			enabled, _ := misc.MapLookup(
+				conn.Config, "source", "syncSettings", "errorDetailsConfig", "enabled",
+			).(bool)
+			index[connectionKey{sourceID: conn.SourceID, destinationID: conn.DestinationID}] = enabled
+		}
+	}
+	d.connectionsMu.Lock()
+	d.connections = index
+	d.connectionsMu.Unlock()
+	d.configLoadedOnce.Do(func() { close(d.configLoaded) })
+}
+
+// connectionEnabled resolves the setting for one connection. It fails closed: an
+// unknown connection, a missing path or a non-boolean value all resolve to false.
+func (d *syncSettingDelegate) connectionEnabled(sourceID, destinationID string) bool {
+	d.connectionsMu.RLock()
+	defer d.connectionsMu.RUnlock()
+	return d.connections[connectionKey{sourceID: sourceID, destinationID: destinationID}]
+}
+
+// configWaitTimeout bounds the startup-window wait in awaitConfig. The caller's context
+// belongs to a router status-update transaction and can be minutes long, which is far
+// too long to hold that path; failing after a few seconds turns the wait into a normal
+// batch error that the router retries, by which time the config has landed.
+const configWaitTimeout = 30 * time.Second
+
+// awaitConfig blocks until the first backend-config push has been indexed.
+//
+// Only the pin-computing path calls it, and only on a genuine cache miss. Without it a
+// failed record arriving in the window between construction and the first push would
+// resolve against an empty index and pin the run to false for the whole retention -
+// permanently wrong, with nothing to indicate it. After the first push this is a read
+// from a closed channel.
+func (d *syncSettingDelegate) awaitConfig(ctx context.Context) error {
+	select {
+	case <-d.configLoaded:
+		return nil
+	default:
+	}
+	ctx, cancel := context.WithTimeout(ctx, configWaitTimeout)
+	defer cancel()
+	select {
+	case <-d.configLoaded:
+		return nil
+	case <-ctx.Done():
+		return fmt.Errorf("waiting for the first backend config push: %w", ctx.Err())
+	}
+}
+
+// staticSyncSettingDelegate answers every question the same way.
+//
+// The SyncSettingDelegate interface takes an unexported type on purpose, so the only
+// real implementation is the one in this package; this is the stand-in for the two
+// cases that do not need it - tests, and the collectors that never collect failed
+// records at all.
+type staticSyncSettingDelegate struct {
+	errorResponse string
+	err           error
+}
+
+// NewStaticSyncSettingDelegate returns a SyncSettingDelegate that always answers with
+// errorResponse and err. Intended for tests.
+func NewStaticSyncSettingDelegate(errorResponse string, err error) SyncSettingDelegate {
+	return &staticSyncSettingDelegate{errorResponse: errorResponse, err: err}
+}
+
+// newUnsupportedSyncSettingDelegate returns the delegate every collector starts with
+// until WithSyncSettingDelegate replaces it.
+//
+// The gateway's collectors and the processor's only ever report Stats, so requiring
+// them to name a delegate would have meant either a compile-time parameter at every
+// call site or a real component - a database connection, a table and a cleanup routine
+// - in processes (the gateway-only pod) that have no use for any of it. Instead they
+// build a collector the ordinary way and get this.
+//
+// It fails loud rather than quiet: if such a collector ever does reach
+// CollectFailedRecords, the very first aborted rETL record returns this error, naming
+// the component, and the strict propagation path aborts the batch - instead of
+// publishing durable failed records with a silently empty error_response.
+func newUnsupportedSyncSettingDelegate(component string) SyncSettingDelegate {
+	return &staticSyncSettingDelegate{
+		err: fmt.Errorf(
+			"the %q stats collector was built without a sync setting delegate and cannot collect failed records",
+			component,
+		),
+	}
+}
+
+func (s *staticSyncSettingDelegate) GetErrorResponse(_ context.Context, _ statKey, _ *jobsdb.JobStatusT) (string, error) {
+	return s.errorResponse, s.err
 }
